restructured folders to seperate their concerns
