+++ frontend/src/components/App.tsx
@@ -2,12 +2,10 @@ import React from 'react'
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 import useMediaQuery from '@mui/material/useMediaQuery'
 import Homepage from './Homepage'
-import Planner from './Planner'
-import AddProgram from './AddProgram'
-import WorkoutList from './workout/WorkoutList'
+import Planner from './program/Planner'
+import { AddProgram, ProgramDetails } from './program/components'
+import { WorkoutList, WorkoutDetails } from './workout/components'
 import GlobalStyle from '../styles/GlobalStyles'
-import ProgramDetails from './ProgramDetails'
-import WorkoutDetails from './workout/WorkoutDetails'
 import NotFound from './NotFound'
 import ProtectedRoute from './auth/ProtectedRoute'
 import NotMobile from './NotMobile'
