+++ frontend/src/components/program/components/index.ts
@@ -1,0 +1,5 @@
+import { AddProgram } from './AddProgram'
+import { ProgramDetails } from './ProgramDetails'
+import { ProgramList } from './ProgramList'
+
+export { AddProgram, ProgramDetails, ProgramList }
