+++ frontend/src/components/workout/components/index.ts
@@ -1,0 +1,5 @@
+import { WorkoutHistoryTableRow } from './WorkoutHistoryTableRow'
+import { WorkoutDetails } from './WorkoutDetails'
+import { WorkoutList } from './WorkoutList'
+
+export { WorkoutDetails, WorkoutHistoryTableRow, WorkoutList }
